refac[recipe]: simplify code by removing loops and utilising atomic transactions
`transaction.atomic()` is a better option here as it prevents the recipe from being
saved if there are any issues with the inlineformsets. Else we get a saved recipe
without any ingredients/instructions, and then are unable to save again due to the
name and slug needing to be unique.

I'm also using get_context_data to help reduce the number of lines required for the
create and update views and streamline the code.

Ref: https://medium.com/@adandan01/django-inline-formsets-example-mybook-420cc4b6225d